+++ src/utility/domElements.js
@@ -1,0 +1,25 @@
+const description = document.querySelector('.description');
+const options = document.querySelector('.options');
+const picture = document.querySelector('.illustration-holder');
+const newGame = document.querySelector('.new-game');
+const menu = document.querySelector('.menu');
+const gameplayUI = document.querySelector('.game-holder');
+const aboutProject = document.querySelector('.about-project');
+const menuButtons = document.querySelector('.menu-buttons');
+const projectInformation = document.querySelector('.project-information');
+const backButton = document.querySelector('.backToMenu');
+const health = document.querySelector('.health');
+const fatigue = document.querySelector('.fatigue');
+const hunger = document.querySelector('.hunger');
+const thirst = document.querySelector('.thirst');
+const inventoryButton = document.querySelector('.inventory-button');
+const skillsButton = document.querySelector('.skills-button');
+const descriptionHolder = document.querySelector('.description-holder');
+const inventoryHolder = document.querySelector('.inventory-holder');
+const skillsHolder = document.querySelector('.skills-holder');
+
+export {
+    description, options, picture, newGame, menu, gameplayUI, aboutProject, menuButtons, projectInformation,
+    backButton, health, fatigue, hunger, thirst, inventoryButton, skillsButton, descriptionHolder, inventoryHolder,
+    skillsHolder
+}
