+++ src/utility/renderScene.js
@@ -2,6 +2,7 @@ import {scenes} from "../content/scenes";
 import {renderOption} from "./renderOption";
 import {renderPicture} from "./renderPicture";
 import {killChildren} from "./killChildren";
+import {description} from "./domElements";
 
 const renderScene = function (id){
     let scene = scenes[id];
